#1540 REST API split: pkg/api/handlers.go + metrics.go → sibling files
Closes #1540.

Pure code motion. Splits two pkg/api monoliths into sibling
pkg/api/<aspect>.go files in the same flat package per parent
wave-1 instructions (no underscore-prefix api_sessions.go-style
filenames; no subdirectory promotion at this step).

pkg/api/handlers.go (2223 LOC, 67 fns) split into 13 files:
- api.go            JSON/HTTP helpers + apiRuntimeDataPlane + queryInt
                    + queryUint16 + allInterfaceNames + applyResult
- health.go         healthHandler, statusHandler
- stats.go          globalStatsHandler, ifaceStatsHandler,
                    zoneStatsHandler, clearCountersHandler
- security.go       zonesHandler, policiesHandler, screenHandler,
                    eventsHandler, matchPoliciesHandler, match* helpers,
                    policyActionStr, screenChecks
- nat.go            natSourceHandler, natDestHandler,
                    natPoolStatsHandler, natRuleStatsHandler
- interfaces.go     interfacesHandler, interfacesDetailHandler,
                    writeInterfacesTerse, writeInterfacesDetail
- routing.go        routesHandler, ospfHandler, bgpHandler
- dhcp.go           dhcpLeasesHandler, dhcpIdentifiersHandler,
                    clearDHCPIdentifiersHandler
- ipsec.go          ipsecSAHandler
- vrrp.go           vrrpHandler
- system.go         systemInfoHandler, systemBuffersHandler,
                    systemActionHandler, pingHandler, tracerouteHandler
- config.go         19 config-mode handlers + configCommitResponse
                    + commitResponseFromConfig
- show_text.go      showTextHandler (~300 LOC dispatcher)
- sessions.go       (renamed from handlers_sessions.go) +
                    protoName moved in from api.go

pkg/api/metrics.go (2033 LOC, 33 fns) split into 7 files:
- metrics.go              slim core: xpfCollector struct + Describe
                          + Collect + emitHistogram + bucketUpperBoundNs
                          + policyCounterID (354 LOC, was 2033)
- metrics_descriptors.go  newCollector (~600 LOC descriptor block)
- metrics_userspace.go    collectUserspaceStatus + 17 emit* methods
                          consuming dpuserspace.ProcessStatus
- metrics_counters.go     collectGlobalCounters, collectInterfaceCounters,
                          collectZoneCounters, collectPolicyCounters,
                          collectFilterCounters
- metrics_sessions.go     collectSessionGauges
- metrics_nat.go          collectNATPoolMetrics
- metrics_system.go       collectDHCPMetrics, collectSystemMetrics,
                          parseMemInfoKB

Plan-review: Codex + AGY both returned PLAN-NEEDS-MINOR on plan v1
(commit 5ed3fdd3). All required minors folded into implementation:
- config_handlers.go renamed to config.go (consistency with nat.go,
  routing.go, dhcp.go, vrrp.go, ipsec.go).
- newCollector ~600 LOC moved out of metrics.go to metrics_descriptors.go
  so metrics.go is truly slim.
- parseMemInfoKB moved to metrics_system.go (only caller is
  collectSystemMetrics).
- api.go tightened: policyActionStr + screenChecks moved to security.go,
  protoName moved to sessions.go, configCommitResponse +
  commitResponseFromConfig moved to config.go. api.go now hosts only
  apiRuntimeDataPlane + writeJSON/writeOK/writeError + queryInt +
  queryUint16 + allInterfaceNames + applyResult.

Validation:
- go build ./pkg/api/... clean
- go test ./pkg/api/... ./pkg/grpcapi/... ./pkg/cli/... all pass
- go test ./... full suite passes
- pkg/dataplane/retirement_boundary_canary_test.go allowlist updated
  to enumerate the new pkg/api/*.go files that still import the legacy
  dataplane bridge (no new imports; same legacy surface, more files).
- docs/pr/1373-retire-ebpf-dataplane/README.md retirement table updated
  to mirror the new allowlist entries.

No behavioral change. No signature change. No exported API change.
server.go, sse.go, auth.go, types.go, all *_test.go untouched.
