+++ docs/pr/1540-rest-api-split/reviewer-ids.md
@@ -8,9 +8,8 @@ results by id rather than re-dispatching.
 
 | Round | Reviewer | Task ID | Verdict | Notes |
 |-------|----------|---------|---------|-------|
-| 1     | Codex    | TBD     | TBD     |       |
-| 1     | Gemini   | TBD     | TBD     |       |
-| 1     | AGY      | TBD     | TBD     |       |
+| 1     | Codex    | (foreground task on 5ed3fdd3) | PLAN-NEEDS-MINOR | 4 items: config_handlers.go→config.go, newCollector→metrics_descriptors.go + parseMemInfoKB→metrics_system.go, tighten api.go scope (move policyActionStr/screenChecks/protoName/configCommitResponse out), atomic-move + build cadence. All folded. |
+| 1     | AGY      | review-mpmuqzeo-dgl1z7 | PLAN-NEEDS-MINOR | 5 items: config_handlers.go→config.go, match* helpers in security.go (already), newCollector→metrics_descriptors.go, parseMemInfoKB→metrics_system.go, per-file build cadence. All folded. |
 
 ## Implementation review
