+++ pkg/api/metrics_system.go
@@ -1,0 +1,113 @@
+package api
+
+import (
+	"bufio"
+	"os"
+	"runtime"
+	"strconv"
+	"strings"
+	"time"
+
+	"github.com/prometheus/client_golang/prometheus"
+)
+
+func (c *xpfCollector) collectDHCPMetrics(ch chan<- prometheus.Metric) {
+	if c.srv.dhcp == nil {
+		return
+	}
+	leases := c.srv.dhcp.Leases()
+	var inet, inet6 int
+	for _, l := range leases {
+		if l.Family == 6 {
+			inet6++
+		} else {
+			inet++
+		}
+	}
+	ch <- prometheus.MustNewConstMetric(c.dhcpLeasesActive, prometheus.GaugeValue,
+		float64(inet), "inet")
+	ch <- prometheus.MustNewConstMetric(c.dhcpLeasesActive, prometheus.GaugeValue,
+		float64(inet6), "inet6")
+}
+
+func (c *xpfCollector) collectSystemMetrics(ch chan<- prometheus.Metric) {
+	// Daemon uptime
+	ch <- prometheus.MustNewConstMetric(c.daemonUptime, prometheus.GaugeValue,
+		time.Since(c.srv.startTime).Seconds())
+
+	// Daemon RSS from /proc/self/statm (field 1 = RSS in pages)
+	if data, err := os.ReadFile("/proc/self/statm"); err == nil {
+		fields := strings.Fields(string(data))
+		if len(fields) >= 2 {
+			if rssPages, err := strconv.ParseUint(fields[1], 10, 64); err == nil {
+				ch <- prometheus.MustNewConstMetric(c.daemonMemRSS, prometheus.GaugeValue,
+					float64(rssPages)*float64(os.Getpagesize()))
+			}
+		}
+	}
+
+	// System memory from /proc/meminfo
+	if f, err := os.Open("/proc/meminfo"); err == nil {
+		defer f.Close()
+		scanner := bufio.NewScanner(f)
+		for scanner.Scan() {
+			line := scanner.Text()
+			if strings.HasPrefix(line, "MemTotal:") {
+				if v := parseMemInfoKB(line); v > 0 {
+					ch <- prometheus.MustNewConstMetric(c.sysMemTotal, prometheus.GaugeValue, float64(v)*1024)
+				}
+			} else if strings.HasPrefix(line, "MemAvailable:") {
+				if v := parseMemInfoKB(line); v > 0 {
+					ch <- prometheus.MustNewConstMetric(c.sysMemAvail, prometheus.GaugeValue, float64(v)*1024)
+				}
+			}
+		}
+	}
+
+	// CPU usage from /proc/stat (instantaneous snapshot)
+	if f, err := os.Open("/proc/stat"); err == nil {
+		defer f.Close()
+		scanner := bufio.NewScanner(f)
+		if scanner.Scan() {
+			line := scanner.Text()
+			if strings.HasPrefix(line, "cpu ") {
+				fields := strings.Fields(line)
+				// fields: cpu user nice system idle iowait irq softirq steal
+				if len(fields) >= 5 {
+					user, _ := strconv.ParseFloat(fields[1], 64)
+					nice, _ := strconv.ParseFloat(fields[2], 64)
+					system, _ := strconv.ParseFloat(fields[3], 64)
+					idle, _ := strconv.ParseFloat(fields[4], 64)
+					iowait := 0.0
+					if len(fields) >= 6 {
+						iowait, _ = strconv.ParseFloat(fields[5], 64)
+					}
+					total := user + nice + system + idle + iowait
+					if len(fields) >= 9 {
+						irq, _ := strconv.ParseFloat(fields[6], 64)
+						softirq, _ := strconv.ParseFloat(fields[7], 64)
+						steal, _ := strconv.ParseFloat(fields[8], 64)
+						total += irq + softirq + steal
+					}
+					cpus := float64(runtime.NumCPU())
+					if total > 0 && cpus > 0 {
+						ch <- prometheus.MustNewConstMetric(c.sysCPUUser, prometheus.GaugeValue,
+							(user+nice)/total*100*cpus)
+						ch <- prometheus.MustNewConstMetric(c.sysCPUSystem, prometheus.GaugeValue,
+							system/total*100*cpus)
+					}
+				}
+			}
+		}
+	}
+}
+
+// parseMemInfoKB extracts the numeric kB value from a /proc/meminfo line.
+func parseMemInfoKB(line string) uint64 {
+	fields := strings.Fields(line)
+	if len(fields) < 2 {
+		return 0
+	}
+	v, _ := strconv.ParseUint(fields[1], 10, 64)
+	return v
+}
